another rewrite

+++ include/chaos/enc.h
@@ -2,7 +2,8 @@
 #define ENC_H
 
 #include <chaos/gen.h>
+#include <stddef.h>
 
-void cipher(struct gen *g, uint8_t *buf, uint64_t len);
+void cipher(struct gen *g, uint8_t *buf, size_t len);
 
 #endif
